% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{coef.york}
\alias{coef.york}
\title{York Regression coefficients}
\usage{
\method{coef}{york}(model)
}
\arguments{
\item{model}{York regression model, as output by `york()`}

\item{data}{Data matrix. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0.}
}
\value{
Estimates and standard errors from the fitted York regression model
}
\description{
Coefficient methods to extract York regression model parameters
}
\details{
...
}
