% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{fitted.york}
\alias{fitted.york}
\title{York Regression predicted values}
\usage{
\method{fitted}{york}(object, ...)
}
\arguments{
\item{object}{York regression model, as output by `york()`}

\item{...}{Must include data matrix named `data`. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0. All other arguments ignored.}
}
\value{
Data frame of predicted values for each observation with columns:
}
\description{
Fitted values method for use with york regression output
}
\details{
Wrapper for more standard interface to `pred.york()`
}
