% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{loss_per_obs.york}
\alias{loss_per_obs.york}
\title{York Regression loss}
\usage{
loss_per_obs.york(data, model = NULL, intercept = NA, slope = NA,
  func = "line")
}
\arguments{
\item{data}{Data matrix. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0.}

\item{model}{York regression model, as output by `york()`. Alternatively, can specify model by providing `slope` and `intercept` values.}

\item{intercept}{Scalar, York regression model intercept.}

\item{slope}{Scalar, York regression model slope.}

\item{func}{Which loss function to report. Either "line" or "predicted_values". See Details.}
}
\value{
Data frame of residuals for each observation with columns:
}
\description{
Residuals and loss function values for each observation in a fitted York regression
}
\details{
...
}
