% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{pred.york}
\alias{pred.york}
\title{York Regression predicted values}
\usage{
pred.york(data, model = NULL, intercept = NA, slope = NA)
}
\arguments{
\item{data}{Data matrix. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0.}

\item{model}{York regression model, as output by `york()`. Alternatively, can specify model by providing `slope` and `intercept` values.}

\item{intercept}{Scalar, York regression model intercept.}

\item{slope}{Scalar, York regression model slope.}
}
\value{
Data frame of predicted values for each observation with columns:
}
\description{
Fitted values method for use with york regression output
}
\details{
...
}
