% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{residuals.york}
\alias{residuals.york}
\title{York Regression residuals}
\usage{
\method{residuals}{york}(model, data)
}
\arguments{
\item{model}{York regression model, as output by `york()`}

\item{data}{Data matrix. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0.}
}
\value{
Data frame of residuals for each observation with columns:
\describe{

  \item{\code{xresid}}{difference between X and predicted X value}

  \item{\code{yresid}}{difference between Y and predicted Y value}

  \item{\code{loss}}{weighted misfit}

}
}
\description{
Residuals method for use with york regression output
}
\details{
...
}
