% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{york}
\alias{york}
\title{York Regression}
\usage{
york(dat, intercept = 0, method = "LS", tol = 1e-15, maxit = 50)
}
\arguments{
\item{dat}{Data matrix. Assumed to be a matrix with columns containing (in order):
X, standard error of X, Y, standard error of Y, and residual correlation.
The correlation column may be omitted, in which case the correlation is
assumed to be 0.}

\item{intercept}{Regression intercept. If 'NA' intercept is freely estimated, otherwise fixed to the given value.}

\item{method}{Method for computing standard errors. Can be one of
"LS" (original least squares SE from York) or "ML" (Titterington).}

\item{tol}{Tolerance for evaluating convergence, based on relative change in the slope.}

\item{maxit}{Maximum number of iterations}
}
\value{
List with elements:
\describe{

  \item{\code{converge}}{list indicating whether the regression fit converged and the tolerance criteria used}

  \item{\code{a}}{fitted intercept and standard error}

  \item{\code{b}}{fitted slope and standard error}

  \item{\code{cov.ab}}{covariance of the slope and intercept estimates}

  \item{\code{chi2}}{goodness of fit statistic}

  \item{\code{df}}{degrees of freedom}

  \item{\code{mswd}}{weighted mean squared residual}

  \item{\code{p.value}}{test of goodness of fit}

}
}
\description{
York Regression
}
\examples{

n <- 50
xt <- rnorm(n)
xs <- runif(n,.005,.05)

yt <- sqrt(.9)*xt + sqrt(.1)*rnorm(n)
ys <- runif(n,.005,.05)

rr <- .5

rand_biv <- function(i){
  ss <- matrix(c(xs[i]^2, rr*xs[i]*ys[i], rr*xs[i]*ys[i], ys[i]), 2, 2)
  ee <- MASS::mvrnorm(1, mu=c(xt[i],yt[i]), Sigma=ss)
}

eo <- t(sapply(1:n, rand_biv))

xo <- xt + eo[,1]
yo <- yt + eo[,2]

yo <- sign(xo)*yo
xo <- sign(xo)*xo

dat <- data.frame(xo,xs,yo,ys,rep(rr,n))

IsoplotR::york(dat)

york(dat)

york(dat, intercept=0)

york(dat, method="ML")

}
\references{
York, D., 1968. Least squares fitting of a straight line with
correlated errors. Earth and Planetary Science Letters, 5, pp.320-324.

Titterington, D.M. and Halliday, A.N., 1979. On the fitting of
parallel isochrons and the method of maximum likelihood. Chemical
Geology, 26(3), pp.183-195.

York, D., et al., 2004. Unified equations for the slope,
intercept, and standard errors of the best straight line. American
Journal of Physics 72.3, pp.367-375.

Mahon, K.I., 1996. The new “York” regression: application of an
improved statistical method to geochemistry. International Geology
Review, 38(4), pp. 293-303.
}
