% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/york.R
\name{coef.york}
\alias{coef.york}
\title{York Regression coefficients}
\usage{
\method{coef}{york}(object, ...)
}
\arguments{
\item{object}{York regression model, as output by `york()`}

\item{...}{Ignored}
}
\value{
Estimates and standard errors from the fitted York regression model
}
\description{
Coefficient methods to extract York regression model parameters
}
\details{
...
}
